Refactor config parsing
"burst" is internally now a time interval, but can be configured either
as an interval or as a number of requests (in which case it will be
multiplied by the domain interval).

interval is still an interval, but can be configured either as an
interval or as a rate (e.g. "3/sec").